+++ src/components/flyout/index.js
@@ -1,0 +1,15 @@
+export {
+  EuiFlyout,
+} from './flyout';
+
+export {
+  EuiFlyoutBody,
+} from './flyout_body';
+
+export {
+  EuiFlyoutFooter,
+} from './flyout_footer';
+
+export {
+  EuiFlyoutHeader,
+} from './flyout_header';