+++ src/components/overlay_mask/index.js
@@ -1,0 +1,3 @@
+export {
+  EuiOverlayMask,
+} from './overlay_mask';